requirements.txt
# ===== Deep Learning & Computer Vision =====
opencv-python==4.9.0.80
mediapipe==0.10.11
tensorflow==2.16.1
numpy==1.26.4
scipy==1.13.0
Pillow==10.3.0
matplotlib==3.8.4
scikit-learn==1.4.2

# ===== Backend =====
flask==3.0.3
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-limiter==3.5.1
pymongo==4.7.1
bcrypt==4.1.3
python-dotenv==1.0.1
marshmallow==3.21.1

# ===== IoT & Simulation =====
requests==2.31.0
paho-mqtt==2.0.0
pygame==2.5.2

# ===== NLP =====
nltk==3.8.1

# ===== Testing =====
pytest==8.1.1
pytest-flask==1.3.0

# ===== Utilities =====
imutils==0.5.4
